move utils to lib
