+++ src/pages/_app.tsx
@@ -2,8 +2,8 @@
 import type { Session } from 'next-auth'
 import { SessionProvider } from 'next-auth/react'
 import type { AppType } from 'next/app'
+import { trpc } from '../lib/trpc'
 import '../styles/globals.css'
-import { trpc } from '../utils/trpc'
 
 const MyApp: AppType<{ session: Session | null }> = ({
   Component,
