+++ src/pages/habits/index.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next'
 import DashboardHabit from '../../components/dashboardHabit'
-import { trpc } from '../../utils/trpc'
+import { trpc } from '../../lib/trpc'
 
 const Habits: NextPage = () => {
   const { data: habitData } = trpc.habit.getUserHabits.useQuery()
